# backend/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"TaskStatus": {
				"enum": [
					"pending",
					"in_progress",
					"completed",
					"cancelled"
				],
				"type": "string"
			},
			"TaskPriority": {
				"enum": [
					"low",
					"medium",
					"high",
					"urgent"
				],
				"type": "string"
			},
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"User": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"password",
					"role",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Task": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/TaskStatus"
					},
					"priority": {
						"$ref": "#/components/schemas/TaskPriority"
					},
					"dueDate": {
						"type": "string",
						"format": "date-time"
					},
					"completedAt": {
						"type": "string",
						"format": "date-time"
					},
					"recurrenceRule": {
						"type": "string"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"metadata": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					},
					"userId": {
						"type": "string"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"projectId": {
						"type": "string"
					},
					"project": {
						"$ref": "#/components/schemas/Project"
					},
					"assignedToId": {
						"type": "string"
					},
					"assignedTo": {
						"$ref": "#/components/schemas/User"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"status",
					"priority",
					"userId",
					"user",
					"order",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Project": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"icon": {
						"type": "string"
					},
					"isShared": {
						"type": "boolean"
					},
					"isArchived": {
						"type": "boolean"
					},
					"metadata": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					},
					"ownerId": {
						"type": "string"
					},
					"owner": {
						"$ref": "#/components/schemas/User"
					},
					"tasks": {
						"items": {
							"$ref": "#/components/schemas/Task"
						},
						"type": "array"
					},
					"members": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"color",
					"isShared",
					"isArchived",
					"ownerId",
					"owner",
					"tasks",
					"order",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_null_": {
				"description": "Standard API Response wrapper\nUsed by all TSOA controllers",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"timestamp": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_Project_": {
				"description": "Standard API Response wrapper\nUsed by all TSOA controllers",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/Project"
					},
					"timestamp": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateProjectDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"icon": {
						"type": "string"
					},
					"isShared": {
						"type": "boolean"
					},
					"memberIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResponse_Project-Array_": {
				"description": "Paginated API Response wrapper\nUsed for endpoints that return lists with pagination",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Project"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					},
					"timestamp": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"message",
					"data",
					"pagination",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProjectDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"icon": {
						"type": "string"
					},
					"isShared": {
						"type": "boolean"
					},
					"isArchived": {
						"type": "boolean"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"metadata": {
						"$ref": "#/components/schemas/Record_string.unknown_"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"AddProjectMembersDto": {
				"properties": {
					"userIds": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"userIds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReorderProjectsDto": {
				"properties": {
					"projectOrders": {
						"items": {
							"properties": {
								"order": {
									"type": "number",
									"format": "double"
								},
								"id": {
									"type": "string"
								}
							},
							"required": [
								"order",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"projectOrders"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Todo AI API",
		"version": "1.0.0",
		"description": "Auto-generated OpenAPI specification for Todo AI",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "API Support",
			"email": "support@todoai.com"
		}
	},
	"paths": {
...
		"/api/projects": {
			"post": {
				"operationId": "CreateProject",
				"responses": {
					"201": {
						"description": "Project created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Project_"
								}
							}
						}
					}
				},
				"description": "Create a new project",
				"summary": "Create a project",
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProjectDto"
							},
							"example": {
								"name": "My Project",
								"description": "Project description",
								"color": "#3B82F6",
								"icon": "📁",
								"isShared": false
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetProjects",
				"responses": {
					"200": {
						"description": "Projects retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponse_Project-Array_"
								}
							}
						}
					}
				},
				"description": "Get all projects with filtering and pagination",
				"summary": "List all projects",
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "isArchived",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "isShared",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"name",
								"createdAt",
								"updatedAt",
								"order"
							]
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"ASC",
								"DESC"
							]
						}
					}
				]
			}
		},
		"/api/projects/{id}": {
			"get": {
				"operationId": "GetProjectById",
				"responses": {
					"200": {
						"description": "Project retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Project_"
								}
							}
						}
					}
				},
				"description": "Get a single project by ID",
				"summary": "Get project by ID",
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateProject",
				"responses": {
					"200": {
						"description": "Project updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Project_"
								}
							}
						}
					}
				},
				"description": "Update a project",
				"summary": "Update project",
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProjectDto"
							},
							"example": {
								"name": "Updated Project Name",
								"description": "Updated description",
								"color": "#10B981"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProject",
				"responses": {
					"200": {
						"description": "Project deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Delete a project",
				"summary": "Delete project",
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/projects/{id}/members": {
			"post": {
				"operationId": "AddMembers",
				"responses": {
					"200": {
						"description": "Members added successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Project_"
								}
							}
						}
					}
				},
				"description": "Add members to a project",
				"summary": "Add project members",
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddProjectMembersDto"
							},
							"example": {
								"userIds": [
									"123e4567-e89b-12d3-a456-426614174000"
								]
							}
						}
					}
				}
			}
		},
		"/api/projects/{id}/members/{userId}": {
			"delete": {
				"operationId": "RemoveMember",
				"responses": {
					"200": {
						"description": "Member removed successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Project_"
								}
							}
						}
					}
				},
				"description": "Remove a member from a project",
				"summary": "Remove project member",
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "User ID to remove",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/projects/reorder": {
			"patch": {
				"operationId": "ReorderProjects",
				"responses": {
					"200": {
						"description": "Projects reordered successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"description": "Reorder projects",
				"summary": "Reorder projects",
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReorderProjectsDto"
							},
							"example": {
								"projectOrders": [
									{
										"id": "project-id-1",
										"order": 0
									},
									{
										"id": "project-id-2",
										"order": 1
									}
								]
							}
						}
					}
				}
			}
		},
		"/api/projects/{id}/archive": {
			"patch": {
				"operationId": "ToggleArchive",
				"responses": {
					"200": {
						"description": "Project archive status updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_Project_"
								}
							}
						}
					}
				},
				"description": "Archive or unarchive a project",
				"summary": "Toggle project archive status",
				"tags": [
					"Projects"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"isArchived": {
										"type": "boolean"
									}
								},
								"required": [
									"isArchived"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}